Revert " Merge tag 'rolling-lts/hcl-main/6.12.42.1' into product/hcl-main/6.12"

+++ net/ipv4/tcp.c
@@ -1145,7 +1145,7 @@ restart:
 		goto do_error;
 
 	while (msg_data_left(msg)) {
-		int copy = 0;
+		ssize_t copy = 0;
 
 		skb = tcp_write_queue_tail(sk);
 		if (skb)
